refactor: share log container selection
